+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Fixes and Changes
 
+- Modules [8FACE, 8FACEx2](./docs/EightFace.md)
+    - Allow disabling of "long press" for changing the number of active slots (#354)
 - Module [MB](./docs/Mb.md)
     - Fixed crash on exiting Rack's after adding MB (#352)